+++ src/bookshelf/cli.py
@@ -1,11 +1,14 @@
 """
 Bookshelf CLI
 """
+import logging
 import os
 
 import click
+import click_log
 
 cmd_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), "commands"))
+logger = logging.getLogger("bookshelf")
 
 
 class _CLICommands(click.MultiCommand):
@@ -20,13 +23,24 @@ class _CLICommands(click.MultiCommand):
     def get_command(self, ctx, cmd_name):
         try:
             mod = __import__(f"bookshelf.commands.cmd_{cmd_name}", None, None, ["cli"])
-        except ImportError:
-            return None  # pragma: no cover
+        except ImportError:  # pragma: no cover
+            return None
         return mod.cli
 
 
 @click.command(cls=_CLICommands, name="bookshelf")
-def main():
+@click.option("-q", "--quiet", is_flag=True)
+@click_log.simple_verbosity_option(logger)
+@click.pass_context
+def main(ctx, quiet):
     """
     Bookshelf for managing reusable datasets
     """
+    ctx.ensure_object(dict)
+
+    if not logger.hasHandlers():
+        click_log.basic_config(logger)  # pragma: no cover
+
+    logger.setLevel(logging.INFO)
+    if quiet:
+        logger.setLevel(logging.ERROR)